+++ genvalidity-containers/default.nix
@@ -4,7 +4,7 @@
 }:
 mkDerivation {
   pname = "genvalidity-containers";
-  version = "1.0.0.1";
+  version = "1.0.0.2";
   src = ./.;
   libraryHaskellDepends = [
     base containers genvalidity QuickCheck validity validity-containers