+++ genvalidity-containers/src/Data/GenValidity/Containers.hs
@@ -3,6 +3,7 @@ module Data.GenValidity.Containers
   )
 where
 
+import Data.GenValidity.IntMap as Containers
 import Data.GenValidity.Map as Containers
 import Data.GenValidity.Sequence as Containers
 import Data.GenValidity.Set as Containers