+++ genvalidity-containers/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [1.0.0.2] - 2022-08-30
+
+### Added
+
+* `GenValid a => GenValid (IntMap a)`
+
 ## [1.0.0.1] - 2022-08-30
 
 ### Added